sndjasn

+++ examples/progression/1_simple.rs
@@ -16,7 +16,7 @@ use viriformat::dataformat::Filter;
 
 fn main() {
     // hyperparams to fiddle with
-    const hl_size : u32 = 2 * 64 * 32;
+    const hl_size : usize = 2 * 64 * 32;
     let initial_lr = 0.001;
     let final_lr = 0.001_f32.powf(5.0);
     let superbatches = 300;
